fix(server): empty links must not mask tampered chain in fleet audit-verify
Codex P2 on #377: `passing` was derived inside the `for l in links`
loop, so a fleet plan with zero linked repos returned `passing: true`
even when the underlying chain was tampered. Initialise `passing`
from `chain.ok` directly so the aggregate always mirrors the chain.

Also moved the route from a standalone `routes/fleet_audit.rs` into
the audit-themed `routes/audit/mod.rs` (where the other audit verbs
live) to fit under the `convergio-server` 14000-line crate cap.
No semantic difference; same URL, same payload shape, same handler
logic.

Codex P2 on #377.
