+++ crates/convergio-server/src/app.rs
@@ -68,7 +68,6 @@ pub fn router(state: AppState) -> Router {
         .merge(crate::routes::embed::router())
         .merge(crate::routes::fleet::router())
         .merge(crate::routes::fleet_plans::router())
-        .merge(crate::routes::fleet_audit::router())
         .merge(crate::routes::telemetry::router())
         .merge(crate::routes::api_actions::router())
         .merge(crate::routes::gate_preconditions::router())
