+++ crates/convergio-server/tests/e2e_fleet_audit_verify.rs
@@ -8,11 +8,10 @@ use sqlx::Executor as _;
 mod common;
 
 #[tokio::test]
-async fn fleet_audit_verify_clean_then_tamper() {
+async fn fleet_audit_verify_detects_tampering() {
     let (base, pool, _d) = common::boot().await;
     let http = Client::new();
     let durability = Durability::new(pool.clone());
-    // Seed: 2 repos, each with a per-repo plan, linked.
     let create: Value = http
         .post(format!("{base}/v1/fleet/plans"))
         .json(&json!({ "title": "audit fleet", "scope": "fleet" }))
@@ -51,21 +50,20 @@ async fn fleet_audit_verify_clean_then_tamper() {
             .error_for_status()
             .unwrap();
     }
-
-    // Clean: all verdicts ok, aggregate passing.
     let url = format!("{base}/v1/fleet/plans/{fid}/audit-verify");
+
+    // Sanity: clean chain → every verdict pass, aggregate passing.
     let r: Value = http.get(&url).send().await.unwrap().json().await.unwrap();
     assert_eq!(r["passing"].as_bool(), Some(true), "{r}");
     let vs = r["verdicts"].as_array().unwrap();
     assert_eq!(vs.len(), 2);
     for v in vs {
         assert_eq!(v["ok"].as_bool(), Some(true), "{v}");
-        assert!(v["broken_at"].is_null(), "{v}");
         assert!(v["checked"].as_i64().unwrap_or(0) > 0, "{v}");
     }
 
-    // Tamper the tail row's prev_hash. Verifier walks in seq order,
-    // so it must trip at that seq.
+    // Tamper the tail row's prev_hash. Verifier walks in seq order
+    // and trips at that seq.
     let (target,): (i64,) = sqlx::query_as("SELECT MAX(seq) FROM audit_log")
         .fetch_one(pool.inner())
         .await
@@ -74,17 +72,35 @@ async fn fleet_audit_verify_clean_then_tamper() {
         .execute(sqlx::query("UPDATE audit_log SET prev_hash='tampered' WHERE seq=?").bind(target))
         .await
         .unwrap();
-
     let r: Value = http.get(&url).send().await.unwrap().json().await.unwrap();
     assert_eq!(r["passing"].as_bool(), Some(false), "{r}");
     for v in r["verdicts"].as_array().unwrap() {
         assert_eq!(v["ok"].as_bool(), Some(false), "{v}");
         assert_eq!(v["broken_at"].as_i64(), Some(target), "{v}");
     }
-    let resp = http
-        .get(format!("{base}/v1/fleet/plans/does-not-exist/audit-verify"))
+
+    // Codex P2 #377: zero linked repos must not mask the now-tampered
+    // chain. Reuse the existing tampering via a second fleet plan.
+    let u: Value = http
+        .post(format!("{base}/v1/fleet/plans"))
+        .json(&json!({ "title": "unlinked", "scope": "fleet" }))
         .send()
         .await
+        .unwrap()
+        .json()
+        .await
         .unwrap();
-    assert_eq!(resp.status().as_u16(), 404, "unknown plan must 404");
+    let r: Value = http
+        .get(format!(
+            "{base}/v1/fleet/plans/{}/audit-verify",
+            u["id"].as_str().unwrap()
+        ))
+        .send()
+        .await
+        .unwrap()
+        .json()
+        .await
+        .unwrap();
+    assert_eq!(r["passing"].as_bool(), Some(false), "{r}");
+    assert!(r["verdicts"].as_array().unwrap().is_empty());
 }
